chat/requirements.txt
bs4==0.0.1
Werkzeug==2.2.2
openai==1.57.2
requests==2.32.3
Flask==2.2.2
scikit-image==0.19.3
scikit-learn==1.6.0
bs4==0.0.1
Pillow==9.2.0
